Create xarg-rsync.py
This script works by first getting a list of all the files in the source directory. It then uses xargs to pass the list of files to rsync. This speeds up the copy because rsync can then read the list of files from stdin instead of having to open each file individually. The --preserve=mode,ownership option tells rsync to preserve the permissions and ownership of the files being copied.